Element: Remove unused elem.selector logic for error messages
The `elem` parameter to `unsafeInfuse()` can be a DOM node, or a
jQuery object. In either case, we pass it to `$()` to form `$elem`,
which does not copy over internal state. Hence the code used
`elem.selector` and not `$elem.selector`, but this also meant the
eslint rule for no-jquery/no-selector-prop didn't spot it.

Anyway, this has been deprecated in jQuery 1.7 and removed long
ago in jQuery 1.11. I added a unit test in I... to test or fix this
diagnostic value (e.g. for T280728), but alas, it has not been
working and can't really, so we should probably remove it for now.

Perhaps in the future a config option could be created to optionally
allow some kind of diagnostic string to be passed along to give a clue
about the element's origin (e.g. naively config.selector, or something
a bit more generic).

<https://github.com/wikimedia/eslint-plugin-no-jquery/blob/v2.6.0/docs/rules/no-selector-prop.md>

Change-Id: I55a568f8665b35c7f1ef7ff931f8c862a3545cd5