Make use of the PHP operator ?? in a few more places
This might need a closer look for two reasons:

1. Using ?? inside if() is certainly shorter, but is it easier
or harder to read? What do you think?

2. When I replace `!empty( $x ) && $x` with `$x ?? false` I'm
aware this is not really the same.

Before: The code checks if $x is set and contains a non-falsy
value. The result is forced to be bool.

After: Whatever non-null value $x contains, the value is used.
It's not forced to be bool.

I feel like this makes more sense. There is a lot of similar
code in this codebase that behaves the same.

Change-Id: Iba7cf21f74ca5c2601a6fb2ed2e0f7ba2e669c1a