Update somewhat ambiguous docs related to the clear indicator
The initial motivation for this patch is that I tried to add
a clear button to a TextInputWidget, but immediatelly run into
issues that made it feel like the implementation is broken.
Indicators don't do anything by default. This can be confusing
if your motivation is to just use the OOUI widgets without
creating your own subclasses.

This patch also simplifies some related methods. Notably: A
TextInputWidget's type is sanitized much earlier. This allows
the following mixin constructors to use the type without the
need to sanitize it again.

"clear" indicators are removed or replaced with "required" in
demos and some docs to reduce confusion. "required" is the
only build-in indicator that makes sense without click
handlers.

Change-Id: I18cafe9420bf548000b150d29394766a9b3bc487