WikimediaUI theme: Remove unnecessary variable
Obstruse `color-primary--active-hover` has only been used in
SelectFileWidget drop target area when a file is dropped and is fine
to receive active primary treatment to simplify code.
This also changes the color slightly to darker Accent30 which is aimed for 
with blue background for better contrast and “active” state of 
interaction at point.

Change-Id: I79f465aae28dbb0b24cbfc2a517e60ae215c1277