Fix BookletLayout.setPage() emitting events twice
This code is pretty much exclusively used in VisualEditor. Most
notably in the template dialog. This patch fixes two bugs:

1. In case a page with the given name can't be found, the
BookletLayout was left in an inconsistent state. Neither the
previousPage nor the .currentPageName property are touched. But the
.outlineSelectWidget is. This effectively removes the selection on
the left side of the dialog. But the BookletLayout class would still
act as if there is a selection.

This patch makes sure nothing happens when .setPage() is called with
an unknown page name. This is closest to the existing behavior.

2. Every .setPage() call triggers the same "set" event twice.
This is because .setPage() needs to update the selection on the
left side of the dialog. This triggers a "select" event that is
indistinguishable from a user clicking an item on the left side of
the dialog. This triggers .setPage() a second time. This loop will
stop when .currentPageName is finally updated. But this happens to
late. The event already fired twice.

This is how the stack trace for the bad "set" event looks like. To
be read from bottom to top. You can see the duplicate .setPage()
call.
  OO.EventEmitter.emit (with "set")
  OO.ui.BookletLayout.setPage
  OO.ui.BookletLayout.onOutlineSelectWidgetSelect (catching the
    "select" event)
  OO.EventEmitter.emit (with "select")
  OO.ui.SelectWidget.selectItem
  OO.ui.SelectWidget.selectItemByData
  OO.ui.BookletLayout.setPage
  OO.ui.BookletLayout.onStackLayoutFocus

This is the stack trace for the expected "set" event.
  OO.EventEmitter.emit (with "set")
  OO.ui.BookletLayout.setPage
  OO.ui.BookletLayout.onStackLayoutFocus

With this change the bad event is gone.

This code is arranged like this ever since it was written via
Id0fb0aa in 2013. I believe it was not identified as a bug because
it does not have much of an effect, except wasting some time on the
client.

Change-Id: I08dc8f20bc190ebbbf08b933c929ac2c7d2fdf9b