Remove misplaced .selectFirstSelectablePage() calls in BookletLayout
Note how this code was not even useful:
* Let's say there are 100 items. Nothing is selected. You add 1
  item. Personally, I would expect that the new item gets selected.
  But instead the list scrolls all the way up to the 1st item.
* Let's say there are 100 items. The last is selected. You remove it.
  Personally, I would expect that item 99 gets selected. But instead
  the list scrolls all the way up to the 1st item.
* Let's say the list is empty. You add 10 items. Are you sure you
  always want to select the 1st item?

While the call in .addPages() should typically not do anything (there
is no way to "loose" a selection there), the call in .removePages()
causes us a lot of trouble. Not only does it cause bad scrolling
that's hard to work around. It triggers a lot of unnecessary event
chains, especially when a dialog initially opens.

I tried to come up with an algorithm that makes more sense, but
realized this is business logic that depends on information outside
of the OOUI lib. For example, the 1st item might be a template name
(indention level 0), and the next is a template parameter (level 1).
Only the VisualEditor template dialog can know which of the two
should be selected.

There are not many users of this code:
https://codesearch.wmcloud.org/search/?q=new%20OO\.ui\.BookletLayout&files=\.js$
There is a chance this patch changes the behavior of some of these
users. The worst case scenario is that nothing is selected, e.g.
when a dialog initially opens. But this is easy to spot, can't cause
data loss, and is still better than the status quo. I tried to
test this change in VisualEditor and was not even able to spot a
difference.

Bug: T289043
Bug: T290975
Change-Id: Id9cf0867710374628f55c25b6d7ea7a81ad9c1e9