Remove bogus auto-scroll behavior from BookletLayout
This is motivated by our work on the VisualEditor template dialog.

This feature does not work as intended. It's more confusing than
helpful. For example:

* As a user, it's unclear how the scroll target is picked. This
makes the feature unpredictable. Esp. when there are multiple pages
on the right of a dialog, which one of them is used? It appears
like sometimes the top one is picked, sometimes the bottom one,
depending on the scroll direction. (Which is extra confusing when I
accidentally move the mouse 1 pixel in the opposite direction before
it stops.)

* Often the item that is picked as a reference is not even visible,
or only partial visible. 1 pixel is apparently enough to be
considered visible. Again, this makes the feature unpredictable.

* The feature removes all focus. I can't, for example, scroll to
another position, have a look, scroll back and continue what I just
did. I even have a hard time finding my old position. It's not
selected any more and the text cursor is gone.

All this depends a lot on the scroll direction and how big the pages
are on the right side of a dialog.

See T289043#7297679 for more details.

Bug: T289043
Change-Id: I98a1d867ac36f36f1f06e08e846fadee2760275c