Fix StackLayout scrolling to the very top when removing items
The existing behavior is probably "better than nothing", but not
really helpful either. In a long StackLayout it can be very
confusing if it scrolls up to the very top all the time. This is
especially notable when the item you are going to remove from the
StackLayout needs to be selected first (i.e. becomes the
currentItem) before you can hit the delete button. This is what
happens in the template dialog in VisualEditor.

The alternative code I'm proposing here does the same as before,
just selects _another_ item that's not necessarily the first. The
idea is to pick a remaining item that directly follows the removed
currentItem. This way we avoid any scrolling in almost all cases.

Another way to solve this is to always call unsetCurrentItem() when
the currentItem was removed. But I feel like this might break the
assumptions some users have.

Bug: T293635
Change-Id: Iab8f2acf5a9081197cb9e00cb3eb50e9a0ba045c