Improve performance of RequiredElement constructor
I found this while profiling the new VisualEditor template editor,
which potentially contains hundreds of checkboxes. Each checkbox
is potentially a required element and calls this code on
construction time.

The issue: .setRequired() is called twice for every checkbox. One
time as part of .setRequiredElement() and another time directly
after. The first call is pointless.

Note I intentionally structured the code in a way so that calling
.setRequired() before .setRequiredElement() is not possible (it
will fail because `this.$required` isn't set). This is intentional
because doing so would still call .updateRequired() twice on
construction time – which is what I aim to fix.

The same time .setRequiredElement() must be called after
`this.required` is set. Again to avoid calling .updateRequired()
twice on construction time.

Bug: T296335
Change-Id: I83398a72270418f575c819ce652ad40452a17433