Skip meaningless default TextInputWidget validation
The default regex /.*/ matches every string. It's fairly cheap,
but not free either, especially because it eats the entire
string. This adds up in large dialogs with possibly hundreds of
these widgets, e.g. the VisualEditor template dialog.

A very common mistake in JavaScript is that .match() is used
instead of .test(). The later returns a boolean and is cheaper
because of this.

Note that while config.validate can be a string, this.validate
can't.

Bug: T296335
Change-Id: I549759449379f0ac6ebb4f98c796e372c783408c