Make MenuSelectWidget's "filterFromInput" mode easier to use
A MenuSelectWidget is the drop-down you get as part of a
ComboBoxInputWidget. Live examples:
https://doc.wikimedia.org/oojs-ui/master/demos/?page=widgets&theme=wikimediaui&direction=ltr&platform=desktop#demo-section-comboBox

The "filterFromInput" mode is off by default.

When it's enabled you can start typing and the items in the drop-down
below are filtered. That's great. But after you picked one it's now
used as the filter and you will never see the other items again. You
have to manually delete the text. That's confusing.

The idea is as follows:
* The moment you enter the field it's "initial" value is temporarily
  remembered.
* We assume that this value is special and _not_ meant to act as a
  filter.
* The moment you start editing we go back to the original filter
  behavior.

Bug: T295885
Change-Id: Ifa47728a0e165c938f2a70ec5ba67b656aaa0454