Fix SelectWidget ignoring selectable/highlightable/pressable
These methods just assumed that calling .setSelected(),
.setHighlighted() or .setPressed() succeeded. But these methods
don't do anything when the corresponding flag is false.

This shouldn't make much of a difference, if at all. The worst-case
is that `aria-activedescendant` points to an element that's not
actually active, and that events are fired with an inactive
element. If there is code that tries to do anything with these
wrong events is unknown at this point.

Change-Id: I2fa361bc8005fb8a1d932b921202dc0819efebe7