Avoid more code duplication in RequiredElement mixin
a) This gets rid of some remaining bits of code duplication.
Notably the separte .removeAttr() calls.

b) Every function executes the most minimal code. For example:
   * .setRequiredElement() doesn't need to update the indicator,
     but was doing this before.
   * There is no point in updating the indicator from the
     constructor when there is nothing to do anyway.
   * The constructor now also skips both .prop( 'required' ) as
     well as .attr( 'aria-required' ) when it can.

Remember, this mixin is used in a lot of widgets, but it's very
rare to actually make one required. Setting the property
`.required = false` on all non-required elements doesn't make a
difference, as it's the default anyway. It just wastes CPU
cycles.

This is a direct follow-up in addition to the changes already
done in I83398a7.

Bug: T296335
Change-Id: Iddba3ddabb0a23442ddd3e1b50a6f3ea63520f1a