ButtonWidget: Fix .setRel() sometimes not working
As discussed in I608ecdc. Calling the method with the same array
reference multiple times should be a rare edge-case. Since the `if`
is not useful anyway we can as well remove it. There is only one
situation where the `if` can work as expected: when .setRel() is
called with an array and the array didn't change in the meantime.

Properly debouncing this is not trivial as there is no fast(er) way
to check if the two arrays contain the exact same values.

Notes:
* Optimized to not store the empty string as a useless array [ '' ].
  Can't do this when the input is an array, as argued above.
* The feature to undo previously set values is usually not needed in
  PHP, as it only creates widgets in their initial state.

This also fixes some related issues:
* ButtonWidget.test.js was not executed.
* .setRel() in JS silently ignored bad values that are neither a
  string nor an array. This will now fail.

Change-Id: I18d459d54590786ba9d7fce8db75ef8607d7f3f4