Fix and update some potentially misleading JSDoc comments
Notable:

* I think it doesn't make sense to have an HtmlSnippet with no content.
The getter is not documented to return undefined.

* I am not able to 100% fix the documentation of Toolbar.setup(). This
is how a `groups` argument actually looks like:

 ….setup( [
     {
         name: '…',
         include: [ … ],
         …
     },
     …
 ] );

This is an array of objects, each with a name, a type, and any number
of arguments accepted by the ToolbarGroup subclass that corresponds to
the type.

I don't know how to document this so that JSDuck doesn't reject it.
The current documentation is still wrong in so far that the properties
are documented as if they are properties of the array. But they belong
to the array elements.

Change-Id: I050f4a2db13cd040de5ad246c21e963dfd2fb445