Fix StyleLint comments wasting space in compiled .css files
I found about 200 of these strange empty sections in MediaWiki core's
resources/lib/ooui/ folder:

 /* stylelint-disable selector-no-vendor-prefix */
 /* stylelint-enable selector-no-vendor-prefix */

It took me a while to understand where this comes from. It's not
StyleLint. It's not the "concat" task. It's the LESS compiler when it
strips unused mixins, but doesn't understand that the comments belong
to the mixin.

This patch reduces the file size of all dist/*.css files. For example,
dist/oojs-ui-wikimediaui.css shrinks from 270 KiB to 260 KiB.

Change-Id: Ia24ffb514cfbdbdc811693082cdf9d803c1d8bf9