Remove unused machinery from IndexLayout PHP implementation
The PHP code is a little different from JS because there is no need
for IndexLayout.removeTabPanels() or IndexLayout.clearTabPanels() in
PHP. All that needs to be done server-side is to initialize these
widgets once.

The reason the JS implementation keeps a reference from TabPanelLayout
to TabOptionWidget is because of the remove and clear features. As
this doesn't exist in PHP we don't need the reference either.

Change-Id: Id72fa49001c0dc2d953edc1268da2ccc4cb41af2