Fix TagMultiselectWidget.setValue behaving odd in edge cases
For example, when you call `.setValue( 3 )` with a number, what the
code tries to do is to find a `.data` property in this number. It
can't find one and passes `undefined` to the next method. This results
in an empty tag being shown, and .getValue() returning weird stuff as
well.

Let's turn this around. While we could check if a `.data` property
exists, we never want objects to be shown as "[object Object]". That's
why I don't check for `.data` but assume that objects are meant to be
like they are. But other stuff (that's numbers, bools, null and
undefined) can be cast to a string.

Bug: T297386
Change-Id: Ieed6bb29f9c44a112f6f228856011643e318cbbe