icons: Align specific language 'bold*' and 'italic*' icons to guidelines
This is based on https://www.figma.com/file/wmY698HOCwGamNXzdtQ952/Bold-and-Italic-icons---T298211
as pre-official, already agreed on update to the WikimediaUI collection,
based on guideline following 'Noto Sans' usage. Reason it's not in
WikimediaUI collection yet is the short time window for OOUI v0.43.0.
It will be delivered there this week as well.

Bug: T298211
Change-Id: Ib1b1f2ece0daea43e5cc154aac3accffee8b3cf0