Fix #scrollIntoView promise never resolving when called repeatedly
When scrollIntoView was called repeatedly, it cancelled the whole
animation queue by calling .stop(), including our own queued function
that would have resolved the previous scrollIntoView's promise.

Instead of queueing the function, pass it in the 'always' option,
so that it's called even when the queue is cancelled.

MenuSelectWidget uses this promise to re-enable mouse interactivity
after responding to keypresses, which is why this caused T302495.

Bug: T302495
Change-Id: Ibe2de70584b3482eb1402d7e988d5badbee3f547