Use flexbox for MessageDialog horizontal layout
As flexbox will overflow the container when too wide
(unlike a table) we can just check if this.$actions has
overflown, instead of all the individual actions, and
having to worry if each action is actually visible.

This fixes T303972 and T303969, and should also fix T137126,
although there isn't a reproducible test case for the latter
on the task.

Bug: T137126
Bug: T303969
Bug: T303972
Change-Id: I7c0701722d6d84131901d8245c77c1aaa14a8f08