PopupWidget: Fix clipping when the popup is forced to be narrower
When ClippableElement forces the popup to become narrower, it sets
width on the $body element, and expects that to affect its wrapper
$popup. However, PopupWidget set an explicit width on $popup, so this
did not work.

Set the width on $body instead (using #setIdealSize). Only small
changes were needed to adjust for that, this was easier than I
expected.

Also update some hacks in ClippableElement, otherwise clipped popups
gain bogus scrollbars. (I checked the demos in the referenced tasks
and those bugs still exist.)

Bug: T180173
Change-Id: I07212af4f2e44127c8fe9c0b507c60752afd8936