WikimediaUI theme, demos: Unify focus outline for high contrast mode
Using what we've learnt in I004aeec567 to use same `outline` value
`1px solid transparent` for all focusable widgets, in order to
cater to Windows High Contrast mode in case.
Also
- clarifying comments to be outspoken about needed subversion of
  Chrome support
- removing duplicated focus rule in one case
- replacing border variable with correct focus one in one widget.

This is a long-term follow-up to T112727 on Chrome behaviour. Chrome >90
seems to have fixed the false `outline` issue, but as we're talking
Basic support, we need to continue with the extra rules.

Change-Id: I142b43f689229332747e879c6d7630292e7c13c9