FieldLayout: Fix and document the behavior when no label is given
When no label is specified, the layouts take up the same width and
height despite different "align" values. This is weird but users of
the library have come to rely on it.

However, when help text is specified, they behave as if the label was
also specified and use the usual different widths and heights.
Previously the layout would be broken in different ways, as for some
combinations of 'helpInline' and 'align' there would be no space in
the interface to display the help text.

Follow-up to cd4e95b88697c88a3fde5c177e1b703d437a0d65.

Bug: T313238
Change-Id: I0e47d17861a8bdca9327b3300a611e67be94af94