DropdownWidget: Add screen reader support while collapsed
In change a7b7007f84 I changed DropdownWidget so that pressing
arrow keys while the menu is collapsed would change the selected
value, instead of expanding the menu.

I didn't realize that ARIA-based screen reader support was only
provided while it was expanded.

MenuSelectWidget:
* Now has a screen reader mode, where the widget isn't visible, but
  acts mostly as if it was: it handles keyboard events and sets some
  ARIA attributes.

DropdownWidget:
* Now activates MenuSelectWidget screen reader mode on focus,
  rather than hackily passing through some events to its functions.

While in the screen reader mode, the menu is not connected to the
dropdown using 'aria-owns'. It is only needed while the control is
expanded, and keeping it connected was causing the menu options'
labels to become a part of the dropdown's label.

Also, 'aria-activedescendant' is only set after the user starts
changing values or expands the menu, otherwise only the descendant
option's label was being read instead of the dropdown's label.

Bug: T314731
Change-Id: Id754a00288652ac3ead6292adb4ec337986c77c0