Make icons demos page more robust
This helps e.g. when testing this locally, but the "demo-image-list"
task was not run yet and the file demos/dist/image-lists.js is
missing. At the moment the demo page crashes for an unknown reason.
With this tiny change it will appear with empty sections with no
icons. This is easier to understand and to debug, I would argue.

Change-Id: I68b287044714d4f737404efaa6c34d5cc4a351be