Allow minlength attribute to be set in text input based widgets
Allow the minlength attribute to be set for any users of JS widgets
that is or are based on TextInputWidget.js. Also implement this
for the php version TextInputWidget.php. The minlength attribute is
similar to the maxlength attribute (which is already possible to
specify) but instead specifies the minimum length the field must
be.

This attribute is not supported by IE 11, but if the browser
does not support it nothing will break and server-side validation
will be needed as is currently the case. The code used to support
the 'minLength' config used to set this attribute is very similar
to the pre-existing code for setting the maxlength attribute.

Bug: T321439
Change-Id: Ibeeafd5949348b96cc4107a3dd14ef5670e1c0cb