WikimediaUI theme: Remove negative top/bottom margins on dropdown menus
This negative margin was originally intended to tuck the menu's
border-top/-bottom “underneath” the triggering element, like a
ButtonMenuSelect-/DropdownWidget it's attached to.
However, the updated designs do not include this negative margin,
and simply place the input's borders directly next to the menu's
border-top, so that both are visible.

Removing this margin makes the focus outline appear to be the
same on all sides of the input, and fixes a couple of bugs
at different zooms in Chrome.

Also putting `border-radius` all around the menu, now that the
negative margin is gone.

Depends-On: Ic649c7f9f9bd9749aabed4e748c4cdd070630d90
Change-Id: I42dbacbdf23b905770768a5a44cfdf39cd37490e