Fix .addItems() methods silently ignoring non-array input
The issue was technically not introduced by Ib1241f9 but that patch
made it obvious. Turns out some callers call .addItems() not with an
array but a single element, even if this was never documented. And
for some reason this worked for a long time. At least for some of
these classes, specifically users of the GroupElement mixin.

Ib1241f9 broke it when it changed `length !== 0` to `!length`. This
is also true for any non-array input that doesn't have a .length
property. Let's just change it back.

Change-Id: Ic250965cf44ed177b8fb8fe8499febdd80c5581f