Work around recent Firefox bug calculating wrong .scrollheight
It's the best quick-fix I can come up with at the moment. Adding it
here does have the advantage that it will fix everything that relies
on this, not only the VisualEditor template dialog where we found this.
The biggest problem is that this fix will only go live with the next
release of OOUI.

Bug: T317369
Change-Id: I7560cebfab6feccfb43e4628b0db65eeacbc11da