Fix MenuSelectWidget removing unexpected items when tabbing out
This code was added in 2020 via Id1ff30a. Note what the comment
explains: "Was only highlighted, now let's select it. No-op if
already selected." This describes the selectItem() method, not
chooseItem(). The later toggles the selection in multiselect mode.
It's not a "no-op if already selected".

This makes me believe this was a trivial mistake.

I tested this as good as I could. Note this bug can only be
reproduced on widgets that allow multi-selection, which is
relatively rare. Some good examples are:
* Go to Special:ApiSandbox, pick "action=query", and play around
  with the "prop", "list", and "meta" fields. They are all
  multiselect.
* Go to Special:Search and play around with the namespace selection
  on wikis where AdvancedSearch is enabled.

Bug: T53229
Bug: T285252
Change-Id: I13857b3d5f1c01065bc387ba9ce2749044f7cf91