TextInputWidget: Use Custom Elements for #onElementAttach support
It took a few years, but browser vendors regained sense and introduced
a replacement for the deprecated 'DOMNodeInsertedIntoDocument' event.
It requires using custom elements from the Web Components APIs, but
that's still so much better than the MutationObserver workaround.
We can use it now since we dropped support for IE 11.

Previously: 27ae8186cf29ebb (Ief78f208a68c52a2), and
3ddb3603404790f (I681038ea42740f32).

Change-Id: I1d2e66906f0c772d0af724806316a87a26441800