icons: Minimize and fix Wikisource logo
This was the only (remaining) icon with white shapes on top of black
shapes. I merged the shapes to a single one. It should look the exact
same on a white background. Except that the new icon now works on all
backgrounds, not only on white.
Equal change to I77ff0bd4e13325.

Bug: T314661
Bug: T328492
Change-Id: I767319690f7ecbe52b77c2f38625851629aebdb9