Fix SelectWidget not firing "choose" event in multiselect mode
As far as I can see this issue exists ever since multiselect mode
was introduced via I0946d11 in 2019. The `changed` flag is never set
to true in multiselect mode, and the event never fired.

My change Ib8bd3e2 is not to blame, I believe. The event was not
fired before.

I suspect we have never seen this causing actual issues because the
multiselect feature is relatively new, relatively rare, and most code
that works with multiselect widgets relies on the "choose" instead of
the "select" event. Still I find this worth fixing.

Bug: T216780
Change-Id: If7de94d52896ee09e17e05ddacdb3a298118e549