ToggleSwitch: Display checked state correctly in RTL
Iafe42246811a83b changed the ToggleSwitch styles in a way that broke the
display of checked ToggleSwitches in RTL. This was caused by a
translateX( calc( 100% + 1px ) ) rule, which CSSJanus doesn't know how
to flip. To make this work again, move that one pixel out of the
translation into a positioning rule (left: 1px;), so that the translation
can be simplified to translateX( 100% ), which CSSJanus does know how to
flip.

(The translateX( calc( 100% + 1px ) ) rule originated in Codex, where it
does work in RTL because postcss-rtlcss is more sophisticated than
CSSJanus and correctly flips it to calc( -1*( 100% + 1px ) ).)

Bug: T341116
Change-Id: I207d7a874e09129591f08303b9de2a841b358d1d