Fix aria-describedby on MultiSelectWidgets
At least on MultiSelectWidgets the DOM element of the actual input
field is on this input property.

The patch makes sure, that the aria-describedby attribute will be
attached there and not on the outer div of the widget. Otherwise
the description will not get read out when using a screenreader
and focusing the field.

Bug: T342216
Change-Id: Ifdf711d6c9736469920ff4aba55811f8248a2620