Fix slightly misplaced ARIA attributes on PopupButtonWidget
These ARIA attributes have been added via I591acf2 not long ago. We
believe this was added to the wrong element. It should be added to
the $button element (an <a> tag) but is currently added to the
parent <div>. The NVDA screenreader apparently can't find it there.
Moving the attributes makes NVDA read something like "this button
does have more info", which the user can explore then.

Found via T342229.

Bug: T293723
Bug: T342229
Change-Id: I69d812ceb08fac1fb9a7c6be3836d7e37182acd1