WikimediaUI theme: Fix two regressions in background color
Following up to Ie568a56cc56.
On MenuOptionWidget (f.e. in DropdownWidget) and OutlineOptionWidget
when selected, and SelectFileWidget when dropping allowed type
of file onto drop area.
On Tools, we go back to the special color for active hover state.
Note, that the selected/highlighted menu follows the simpler one color
approach in Codex now.

Bug: T350071
Change-Id: If2355f5074871c5579d104309ef0407660c354bb