Clean up SelectWidget.getItemMatcher for empty queries
This is not only done for performance reasons, but mostly to make
it quite a bit more obvious what this code is doing.

Note how the method is called every time the `query` string changes
(called from onDocumentKeyPress). When we know the query string is
empty we can bail out much earlier, always return true (or false
in "exact" mode), and ignore the individual `item` and their
`getMatchText` completely.

Motivated by Ib9ff8f0.

Change-Id: I3710337571adcf5d6a87f7536927d4ead5c06ee5