WikimediaUI theme: Increase margin for PopupWidget label to account for close button
For Popups with long titles / labels, the text of the label can
overlap with the close button on account of the surrounding margin being
to small. Increase the end margin to be the size of the
clickable area of the close button.
Also update the demos to feature an overlong title.

Bug: T355059
Change-Id: I0faf4a0c83ffe3aa35b955e9011fcdb710591564