WikimediaUI Theme: minimum updates for dark mode compatibility
As an alternative to updating the @wikimedia/codex-design-tokens
package to version 1.7.0 (which introduces color tokens which are
"dark-mode ready" insofar as they respect the values of CSS
custom props which have aleady been set on the page), this patch
simply vendors in the Codex color tokens into a "colors.less" file.

Doing this allows OOUI to continue on version 1.3.5 of Codex design
tokens indefinitely. The advantage of this approach is that OOUI
styles will not need to be converted over to be compatible with rem
units, which current and future versions of Codex will use for most
size and space values.

Bug: T364777
Change-Id: I7705c76aede8b3c60d4a6bad176f8b228f811af5