TextInputWidget.validationPatterns: Use cheaper regex pattern for non-empty check
We don't need to validate the entire string. We only need to see if
it contains at least 1 character, and then stop.

This obviously won't make any difference we can actually measure. The
strings are way to short for this to be relevant. Still I believe the
code is more correct like this.

Change-Id: I4341ea11ba087c0458dd3e43c7514ad7c280817f