Ignore clicks on the scrollbar in all document mouseup/down handlers
The intent of these handlers is to treat clicking elsewhere in the
document as moving UI focus away from a widget. However, using the
scrollbar shouldn't count for this.

We detect scrollbar clicks because the browser reports them as having
happened on the documentElement, which is otherwise impossible to
directly click.

Bug: T388302
Change-Id: Ic34c230333a703ec149d693b31231000ae8842aa