DropdownInputWidget: Redo dropdown arrow for compat with MediaWiki dark mode
Instead of the unusual approach of applying an indicator class on
the native `<select>`, just render a real IndicatorWidget, which will
automatically behave correctly in dark mode.

We already use this approach in ComboBoxInputWidget. The
implementation here is a bit different, since in addition to the PHP
version, we want to display it in the JS version on mobile as well.

Bug: T369576
Change-Id: If19c44949e18477aa6543a9db4213fae3a747947