SelectFileInputWidget: fix compareFile
This comparison broke, because File objects are not enumerable. They
also do not have ownProperties. Instead simply compare the base
properties of a File to shallow objects.

This would fail if the file contents are different and all the
other properties are the same, but that is not a very likely scenario.

Follow-up to: I25c2f1399997bc625abd7fd2220679e4b04183d3

Bug: T402623
Change-Id: I2bacc06363b8eb19238d1487b996694f29ec6fe6