Add missing preventDefault to SearchWidget.onQueryKeydown
We found this while working on T403588. In the reference "reuse"
search you can see that the cursor up/down keys do two things: You
navigate the list of search results up/down (as intended) and at
the same time the text cursor in the input field moves left/right.

Bug: T403588
Change-Id: If1f80d181bd99f8f7ac777530b8aa4f3029d2c89