TabIndexedElement: Provide a default setLabelledBy() method
Override Element::setLabelledBy() with a method that sets
'aria-labelledby' on the focusable node instead of the root node.
This is a much better default for anything that can have a tabindex.

The problem rarely came up in practice because most TabIndexedElements
can be labelled with a `<label for>` instead of 'aria-labelledby' (see
the getInputId() method). The new override will only make a difference
when a custom widget calls this method, e.g. MW SelectWithInputWidget.

Bug: T411676
Change-Id: I24735d43dff7cd6d344de5e278f97884d4db0dee